bugfix: PR comments
