+++ lib/app/features/chat/e2ee/providers/e2ee_messages_subscriber.c.dart
@@ -22,7 +22,7 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'e2ee_messages_subscriber.c.g.dart';
 
-@Riverpod(keepAlive: true)
+@riverpod
 class E2eeMessagesSubscriber extends _$E2eeMessagesSubscriber {
   @override
   Stream<void> build() async* {
