+++ lib/app/features/chat/providers/user_chat_relays_provider.c.dart
@@ -3,7 +3,6 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/exceptions/exceptions.dart';
 import 'package:ion/app/features/auth/providers/auth_provider.c.dart';
-import 'package:ion/app/features/chat/e2ee/providers/e2ee_messages_subscriber.c.dart';
 import 'package:ion/app/features/ion_connect/ion_connect.dart';
 import 'package:ion/app/features/ion_connect/model/action_source.dart';
 import 'package:ion/app/features/ion_connect/model/event_reference.c.dart';
@@ -61,7 +60,6 @@ class UserChatRelaysManager extends _$UserChatRelaysManager {
   /// Signs and broadcasts new chat relay list if an update is needed
   ///
   Future<void> sync() async {
-    ref.read(e2eeMessagesSubscriberProvider);
     final pubkey = ref.read(currentPubkeySelectorProvider);
     if (pubkey == null) {
       throw UserMasterPubkeyNotFoundException();
